+++ test/components/InputTime.spec.ts
@@ -23,6 +23,9 @@ describe('InputTime', () => {
     ['with modelValue', { props: { modelValue: new Time(12, 30) } }],
     ['with default value', { props: { defaultValue: new Time(12, 30) } }],
     ['with placeholder', { props: { placeholder: new Time(12, 30) } }],
+    ['with range', { props: { range: true } }],
+    ['with range and modelValue', { props: { range: true, modelValue: { start: new Time(12, 30), end: new Time(14, 30) } } }],
+    ['with range and defaultValue', { props: { range: true, defaultValue: { start: new Time(12, 30), end: new Time(14, 30) } } }],
     ['with disabled', { props: { disabled: true, modelValue: new Time(12, 30) } }],
     ['with required', { props: { required: true, modelValue: new Time(12, 30) } }],
     ['with readonly', { props: { readonly: true, modelValue: new Time(12, 30) } }],
@@ -35,19 +38,31 @@ describe('InputTime', () => {
     ['with icon', { props: { icon: SignIcon } }],
     ['with leadingIcon', { props: { leadingIcon: Cross30Icon } }],
     ['with trailingIcon', { props: { trailingIcon: SignIcon } }],
+    ['with separatorIcon', { props: { range: true, separatorIcon: Cross30Icon } }],
     ...sizes.map((size: string) => [`with size ${size}`, { props: { size } }]),
     ['with ariaLabel', { attrs: { 'aria-label': 'Aria label' } }],
     ['with as', { props: { as: 'section' } }],
     ['with class', { props: { class: 'max-w-sm' } }],
     ['with b24ui', { props: { b24ui: { base: 'rounded-full' } } }],
     // Slots
-    ['with default slot', { slots: { default: () => 'Default slot' } }]
+    ['with leading slot', { slots: { leading: () => 'Leading slot' } }],
+    ['with default slot', { slots: { default: () => 'Default slot' } }],
+    ['with trailing slot', { slots: { trailing: () => 'Trailing slot' } }],
+    ['with separator slot', { slots: { separator: () => '=' } }]
   ])
 
   describe('emits', () => {
     test('update:modelValue event', async () => {
       const wrapper = await mountSuspended(InputTime)
       const time = new Time(12, 30)
+
+      await wrapper.setValue(time)
+      expect(wrapper.emitted()).toMatchObject({ 'update:modelValue': [[time]] })
+    })
+
+    test('update:modelValue event range', async () => {
+      const wrapper = await mountSuspended(InputTime, { props: { range: true } })
+      const time = { start: new Time(12, 30), end: new Time(14, 30) }
 
       await wrapper.setValue(time)
       expect(wrapper.emitted()).toMatchObject({ 'update:modelValue': [[time]] })