+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * feat(Table): implement row pinning
 * feat(unplugin): add support for prose components
 * feat: add standalone Vue REPL playground
+* feat(InputTime): add `range` prop
 * 
 * fix(Error): support `status` and `statusText` properties
 * fix(ContentSurround): handle RTL mode